Add `Profile` component tests and update `AuthContext`.
Added a test for profile component for all users, and authenticated
user.
Updated components to import useAuth correctly as it's not the default
export.
Added customRender to use in tests as has the AuthContext wrapping
children.
`AuthContext` exports IAuthContext, and itself as default.
`AuthContext` now takes in an initial IAuthContext for easier testing
purposes.